chapter_001/src/test/java/ru/job4j2/array/DefragmentTest.java
package ru.job4j2.array;

import org.junit.Test;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

public class DefragmentTest {
    @Test
    public void notFirstNull() {
        String[] input = {"I", null, "wanna", null, "be", null, "compressed"};
        String[] compressed = Defragment.compress(input);
        String[] expected = {"I", "wanna", "be", "compressed", null, null, null};
        assertThat(compressed, is(expected));
    }

    @Test
    public void firstNull() {
        String[] input = {null, "I", "wanna", null, "be", null, "compressed"};
        String[] compressed = Defragment.compress(input);
        String[] expected = {"I", "wanna", "be", "compressed", null, null, null};
        assertThat(compressed, is(expected));
    }

    @Test
    public void whenAllNullExceptOne() {
        String[] input = {null, null, null, null, "One"};
        String[] compressed = Defragment.compress(input);
        String[] expected = {"One", null, null, null, null};
        assertThat(compressed, is(expected));
    }
}